set up tests
