+++ amgut/lib/data_access/tests/__init__.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python
+from __future__ import division
+
+# -----------------------------------------------------------------------------
+# Copyright (c) 2014--, The American Gut Project Development Team.
+#
+# Distributed under the terms of the BSD 3-clause License.
+#
+# The full license is in the file LICENSE, distributed with this software.
+# -----------------------------------------------------------------------------
